add global command flag to alias add
Works with exisiting aliases and extends the command syntax to capture
the global flag and pass that along to AliasedCommand.

Updates `add_alias_route` to make the decision on whether it should match
a command or message listener.

Not working yet: after restarting process, load_aliases does not register
global aliases, as the state is not persisted.